Revert stubbing of runas functions
All the runas functions were stubbed on builds where the sessiond isn't
built which is the case for all platforms except Linux. This was done
because of 2 new commands that require elf.h which is not present on
MacOSX. However the other commands can be used by the relayd.

Revert this and and only stub the relevant commands when "elf.h" is not
present on the system.

+++ src/common/lttng-elf.c
@@ -56,13 +56,13 @@
 	do {					\
 		switch (sizeof(x)) {		\
 		case 8:				\
-			x = be64toh(x);		\
+			x = be64toh((uint64_t)x);		\
 			break;			\
 		case 4:				\
-			x = be32toh(x);		\
+			x = be32toh((uint32_t)x);		\
 			break;			\
 		case 2:				\
-			x = be16toh(x);		\
+			x = be16toh((uint16_t)x);		\
 			break;			\
 		case 1:				\
 			break;			\
